Set the invoice_reference in the save method
The invoice reference for ProjectPayouts was set in a the project's
post_save signal handler. As far as I can tell, that is the only way
that payouts are created. However the database of onepercent contains
payouts that do not have the invoice_reference being set.

By moving this to the save method, we will know for sure that the
invoice reference is set. As far as I can tell, this should always be
the case.