Make sorl.thumbnail a tenant app
When uploading a file with an identical filename on more then one
tenant, sorl thumbnail generates the same kv cache key for all the
files. This means the file can only be retrieved for one of the tenants.
One way to hit this issue is by using facebook to login for more then 1
tenant. Since facebook send the same filename to each tenant, only one
tenant will show a profile image.

This change makes sure we have a separate kv store for each tenant, so
that we do not have these conflicts