Do not mark payments as created.
This is not a a valid choice in the model. If something goes wrong
during the payout creation, this status will not show, and it will not
be possible to change the status.

Under normal conditions, the status would only be set to "created"
while waiting for the payout app do do the status callback. So in that
situation, it also is not useful,

BB-9613 #resolve