When a create requests creates a user that was created outside of SCIM, turn this request into an update.
When no scim_external_id is set, the user is created outside of the scim
flow. In this case we turn the create in an update, and act as if the
user was created. This will set the external id, and will cause a
uniquness error on the next try.